Origin isolation: a new strategy for window.originIsolationRestricted
In https://chromium-review.googlesource.com/c/chromium/src/+/2243994 I
introduced an implementation for window.originIsolationRestricted which
pipes the isolation state from NavigationRequest to the navigated-to
LocalDOMWindow. However, this does not work for the case of the initial
about:blank, where no navigation is performed. Furthermore, it does not
match the spec, where the Window property just reflects an agent
cluster-wide property.

This CL introduces an alternate approach, more similar to what is done
for self.crossOriginIsolated in
https://chromium-review.googlesource.com/c/chromium/src/+/2247463, which
is another agent cluster-wide value. The origin isolation state is
stored in the renderer process as a process-wide global, as part of the
Agent class. Then the LocalDOMWindow getter can just pick it up from the
surrounding agent, as in the spec.

This fixes the problem with about:blank. HOWEVER, TODO BEFORE MERGING:

- I'm still setting this per-render-process value from the
  NavigationRequest code. This feels wrong, since navigations happen
  many times per process/agent; ideally we'd only want to set this once
  per process/agent. I've tried to make this a bit better by adding a
  DCHECK such that if we ever set the origin isolation state more than
  once, subsequent sets must match the previously-set value. But it'd be
  better to only set it once.

- This causes a crash when data: URLs are involved. The existing test,
  external/wpt/origin-isolation/getter-special-cases/data-url.https.html,
  shows this crash in action. In particular it fails the
  newly-introduced DCHECK. I suspect this is because the data: URL page
  is not getting its own process, despite having an opaque origin.

Bug: 1095653
Change-Id: I8dfa8fc4a4766efc0611d43a255673662c422776