webrtc wpt: fix split.https.html firefox compatibility
The test does not work as intended in Firefox as described in
  https://bugzilla.mozilla.org/show_bug.cgi?id=1643050
since Firefox does not use the same fingerprint/certificate for different
peerconnections on the same page and puts the fingerprint into the SDP sessionpart.
With the change the test connects and gets video on the other end but still
fails in Firefox since it relies on connectionState which is not implemented.

BUG=None

Change-Id: Id6529bcb32f37e657d6f79c8d94a4b69c4d15db3